labs/lab-3-fifa-tools/requirements.txt
# Dependencias para FIFA Agent Plus con Google Search nativo de ADK
google-adk>=0.1.0